Adds "mat_csr_sort_rows".
Note that this function is currently *not* being used.  The reason
for this is that even if a sparse matrix is "sorted" to begin with,
the pre-processing stage of the solving code does not preserve this,
and to complicate the issue, the data vector "x" in the solve
structure is only a pointer.